prepare BaseNeighbors and move its inner SlidinWindow class from skmultiflow/utils
