+++ scikit-multiflow/src/skmultiflow/utils/data_structures.py
@@ -865,135 +865,3 @@ class MOLConfusionMatrix(object):
                ' - total_sum: ' + str(self.get_total_sum()) + \
                ' - total_discordance: ' + str(self.get_total_discordance()) + \
                ' - dtype: ' + str(self.dtype)
-
-
-def InstanceWindow(self, n_features=0, n_targets=1, categorical_list=None,
-                   max_size=1000, dtype=float):   # pragma: no cover
-    warnings.warn("'InstanceWindow' has been replaced by 'SlidingWindow' in v0.5.0 "
-                  "and will be removed in v0.7.0",
-                  category=FutureWarning)
-    return SlidingWindow(window_size=max_size)
-
-
-class SlidingWindow(object):
-    """ Keep a fixed-size sliding window of the most recent data samples.
-
-    Parameters
-    ----------
-
-    window_size: int, optional (default=1000)
-        The window's size.
-
-    Raises
-    ------
-    ValueError
-        If at any moment, a sample with a different number of attributes than
-         those already observed is passed.
-
-    Notes
-    -----
-    It updates its stored samples by the FIFO method, which means
-    that when size limit is reached, old samples are dumped to give
-    place to new samples.
-
-    The internal buffer does not keep order of the stored samples,
-    when the size limit is reached, the older samples are overwritten
-    with new ones (circular buffer).
-
-    """
-
-    def __init__(self, window_size=1000):
-        super().__init__()
-
-        self.window_size = window_size
-        self._n_features = -1
-        self._n_targets = -1
-        self._X_queue = None
-        self._y_queue = None
-        self._is_initialized = False
-
-    def configure(self):
-        self._X_queue = np.zeros((0, self._n_features))
-        self._y_queue = np.zeros((0, self._n_targets))
-        self._is_initialized = True
-
-    def add_sample(self, X, y):
-        """ Add a (single) sample to the sample window.
-
-        X: numpy.ndarray of shape (1, n_features)
-            1D-array of feature for a single sample.
-
-        y: numpy.ndarray of shape (1, n_targets)
-            1D-array of targets for a single sample.
-
-        Raises
-        ------
-        ValueError: If at any moment, a sample with a different number of \
-        attributes than that of the n_attributes parameter is passed, a \
-        ValueError is raised.
-
-        TypeError: If the buffer type is altered by the user, or isn't \
-        correctly initialized, a TypeError may be raised.
-
-        """
-        if not self._is_initialized:
-            self._n_features = get_dimensions(X)[1]
-            self._n_targets = get_dimensions(y)[1]
-            self.configure()
-
-        if self._n_features != get_dimensions(X)[1]:
-            raise ValueError("Inconsistent number of features in X: {}, previously observed {}.".
-                             format(get_dimensions(X)[1], self._n_features))
-
-        if self.size == self.window_size:
-            # Delete oldest sample
-            self._X_queue = np.delete(self._X_queue, 0, axis=0)
-            self._y_queue = np.delete(self._y_queue, 0, axis=0)
-
-        self._X_queue = np.vstack((self._X_queue, X))
-        self._y_queue = np.vstack((self._y_queue, y))
-
-    def delete_oldest_sample(self):
-        """ Delete the oldest sample in the window. """
-        if self.size > 0:
-            self._X_queue = self._X_queue[1:, :]
-            self._y_queue = self._y_queue[1:, :]
-
-    def reset(self):
-        """ Reset the sliding window. """
-        self._n_features = -1
-        self._n_targets = -1
-        self._X_queue = None
-        self._y_queue = None
-        self._is_initialized = False
-
-    @property
-    def features_buffer(self):
-        """ Get the features buffer.
-
-        The shape of the buffer is (window_size, n_features).
-        """
-        return self._X_queue
-
-    @property
-    def targets_buffer(self):
-        """ Get the targets buffer
-
-        The shape of the buffer is (window_size, n_targets).
-        """
-        return self._y_queue
-
-    @property
-    def n_targets(self):
-        """ Get the number of targets. """
-        return self._n_targets
-
-    @property
-    def n_features(self):
-        """ Get the number of features. """
-        return self._n_features
-
-    @property
-    def size(self):
-        """ Get the window size. """
-        return 0 if self._X_queue is None else self._X_queue.shape[0]
